Licencing change
Following on from conversations between Diamond Light Source and the original CCP13 grant holders (John Squire, Trevor Forsyth and Alan Windle) it has been agreed that in order to this repository to be of greatest use to the community it is to be relea
sed under a BSD 3-Clause licence.

This new licence superceedes all previous licencing within this repository, where applicable.